backend/requirements.txt
# Flask Core
Flask==3.0.0
Flask-CORS==4.0.0
python-dotenv==1.0.0

# Database & Data Processing
pandas==2.1.4
numpy==1.26.2

# AI/ML - Time Series Forecasting
prophet==1.1.5
cmdstanpy==1.2.0

# AI/ML - Anomaly Detection
scikit-learn==1.3.2

# HTTP Requests
requests==2.31.0

# Date/Time Utilities
pytz==2023.3
python-dateutil==2.8.2

# Flask Dependencies
Werkzeug==3.0.1
click==8.1.7
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3

# Optional: Development/Testing
pytest==7.4.3
black==23.12.1
